Organize login components into folder
