+++ arches_lingo/src/arches_lingo/App.vue
@@ -5,7 +5,7 @@ import Toast from "primevue/toast";
 
 import { userKey } from "@/arches_lingo/constants.ts";
 import HomePage from "@/arches_lingo/components/HomePage.vue";
-import LoginPage from "@/arches_lingo/components/LoginPage.vue";
+import LoginPage from "@/arches_lingo/components/login/LoginPage.vue";
 
 import type { User } from "@/arches_lingo/types";
